Avoid division between different dtypes

+++ src/evox/algorithms/mo/rvea.py
@@ -128,13 +128,13 @@ class RVEA(Algorithm):
         self.pop = survivor[~nan_mask_survivor]
         self.fit = survivor_fit[~nan_mask_survivor]
 
-        if self.gen % (1 / self.fr) == 0:
+        if self.gen % int(1 / self.fr) == 0:
             self.reference_vector = self._rv_adaptation(survivor_fit)
 
     @trace_impl(_update_pop_and_rv)
     def _trace_update_pop_and_rv(self, survivor: torch.Tensor, survivor_fit: torch.Tensor):
         if_else = TracingCond(self._rv_adaptation, self._no_rv_adaptation)
-        self.reference_vector = if_else.cond(self.gen % (1 / self.fr) == 0, survivor_fit)
+        self.reference_vector = if_else.cond(self.gen % int(1 / self.fr) == 0, survivor_fit)
 
         self.pop = survivor
         self.fit = survivor_fit
